weatherDataMaking/requirements.txt
tkinter
tkinter.ttk
pillow
pandas
json